+++ test/database/XMLModelLoaderTest.cxx
@@ -67,7 +67,7 @@ TEST_CASE("XML database model loader test")
     });
 
     CHECK(result);
-    REQUIRE(models.size() == 12);
+    REQUIRE(models.size() == 14);
     std::set<std::string, std::less<>> languages;
     for (const auto& model : models) {
       languages.insert(model.getLanguage());